% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_description.R
\name{add_description}
\alias{add_description}
\title{Add a description of the variable to the dataframe column'}
\usage{
add_description(col_object, col_name, description_dictionary, ticked = F)
}
\arguments{
\item{col_object}{(column from dataframe) The column from the dataframe which
needs a description}

\item{col_name}{(string) The column name to be found in the dictionary}

\item{description_dictionary}{(dataframe) The data dictionary where the
description of the variable can be found}

\item{ticked}{(boolean) Default: FALSE. If a variable is a 'flag' or not.}
}
\value{
(dataframe)  Returns the column with same attributes as before with
added descriptive label.

 @examples core$sex<-add_description(core$sex,
                                     col_name='sex',
                                     description_dictionary = fields)
}
\description{
`add_description` adds a description (found in the specified data dictionary)
to the column for use in constructing the central dataframe. Adding descriptions
to each variable is an important part of data sharing and reproducibility.
}
