% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.summ_to_string.R
\name{cat.summ_to_string}
\alias{cat.summ_to_string}
\title{Create summaries from categorical data'}
\usage{
cat.summ_to_string(
  categorical.data.name,
  dataframe_object,
  population.list.obj = itt,
  by_arm = T
)
}
\arguments{
\item{categorical.data.name}{(string) Categorical variable which is a column name}

\item{dataframe_object}{(dataframe) A dataframe you wish to summarise}

\item{population.list.obj}{(list) List object of the participant IDs for people
in each arm.}

\item{by_arm}{(boolean) Default TRUE. Whether to provide these summaries by
arm and overall}
}
\value{
(list)  Returns the descriptive summaries in the form of a list
which can be inputted straight into a summary table

 @examples sex_summary<-cont.summ_to_string(core.dataframe,
                                            continuous.data.name = 'sex_demographics')
}
\description{
`cat.summ_to_string` create descriptive summaries for categorical data.
This function is used in the summary table functions. It out puts the categorical
summaries as a list of lines you would see in a baseline summary table e.g.
Male, 54 (96%) etc
}
