% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary_rm_na.R
\name{categorical_summary_rm_na}
\alias{categorical_summary_rm_na}
\title{Returns summary of categorical data '}
\usage{
categorical_summary_rm_na(data)
}
\arguments{
\item{data}{(vector) Vector of categorical data (can include NAs- these will be ignored).}
}
\value{
(list) returns the summary object
}
\description{
`categorical_summary_rm_na`  Returns summary of categorical data
}
\examples{
categorical_summary_rm_na(as.factor(c(rep('cow', 2), rep('sheep',5), rep('pig', 3), rep(NA, 4))))
}
