% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_baseline_table.R
\name{construct_baseline_table}
\alias{construct_baseline_table}
\title{Creates the baseline characteristic table for reporting}
\usage{
construct_baseline_table(
  trial.data,
  var.spec = variable.details.df,
  id_cols = c("screening", "event_name", "event_id"),
  population.list.obj = itt,
  fill_baselineNAs_screening_data = "n"
)
}
\arguments{
\item{trial.data}{(dataframe) the dataframe containing all the trial data -
ideally created by the data_construction function in this package.}

\item{var.spec}{(dataframe) the dataframe within the inputs folder which
specifies which variables to be seen in the baseline table - typically
called variable.details.df}

\item{id_cols}{(vector) the names of the id columns as a character vector.
Default is `c("screening", "event_name", "event_id")`.}

\item{population.list.obj}{(list) a list object of the same length as the
number of arms. Each list element is a character vector containing the
screening numbers of participants in an arm in the relevant analysis
population.}

\item{fill_baselineNAs_screening_data}{('y'/'n')}
}
\value{
(dataframe) returns a dataframe which summarises the baseline characteristics
}
\description{
`construct_baseline_table` returns a dataframe which summarises the
  baseline characteristics per arm and overall specified within the
  AnalysisVariablesDetails.xlsx. With mean, median, sd, IQR, min and max for
  numerical variables, as well as counts of the number of each non-empty item
  in both numeric and categorical variables including each group within the
  categorical variables.
}
