% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_central_dataframe.R
\name{construct_central_dataframe}
\alias{construct_central_dataframe}
\title{Central dataframe construction}
\usage{
construct_central_dataframe(
  analysis_variable_df = .,
  dataframe_list = master,
  visit_df_name = "visit_completion",
  screening_df_name = "identification_log",
  number.arms = N.Arms,
  id_cols = id_cols,
  field_df = fields,
  include.adverse = F,
  blinded = "n"
)
}
\arguments{
\item{analysis_variable_df}{(data frame) should be the data frame in the
Inputs folder. An example can be seen in the template folder.}

\item{dataframe_list}{(list) the master list produced by Read-data.R}

\item{visit_df_name}{(string) This is the name of the PROSPECT csv
file which contains the details of all visits. In some
trials, this is `"visit_completion"` or `"events"`. Default is
`"visit_completion"`.}

\item{screening_df_name}{(string) This is the name of the PROSPECT csv
file which contains the details of all participants screened. In some
trials, this is the "identification_log" or simply "screening_log".}

\item{number.arms}{(integer) the number of arms in the trial.}

\item{id_cols}{(vector) the names of the id columns as a character vector.
Standard is `c("screening", "event_name")` and any additional variables
to be used as an identifier (so will be used when merging dataframes)
should be added AFTER the standard.}

\item{include.adverse}{(boolean) whether to include adverse events as seperate
or as part of the central dataframe. Default is FALSE.}

\item{blinded}{(character) This should be y or n}

\item{field.description.df}{(data frame) the data frame containing the field
labels. Default is `fields`, this is often read in from Fields.csv.}
}
\value{
(list of 2 items: dataframe and list of dataframes) central dataframe
 to be used as a base for analyses or sharing AND the revised master list -
 some variables renamed where duplicated.
}
\description{
`construct_central_dataframe` returns a central data frame which can be used
  as a base dataset for primary analyses or for sharing.
}
\details{
From master list created by Read-data.R, this function will produce a
  dataframe of all participants (from screening onwards) with all the
  variables from the CRFs specified in
  Inputs/AnalysisVariablesDetails.xlsx.

   Any constant characteristics for each participant will be displayed in its
   own column, for example: sex, ethnicity.

  Adverse events can also be included to this dataset when
  include.adverse = TRUE (no need to specify).

  If blinded='y', a dummy randomisation will be implemented for those
  participants with a baseline event. If blinded='n', the
  randomisation allocation will be taken from PROSPECT.

  If there are duplicate variable names, the dataframe names will be appended to
  them e.g. age_demographics
}
\examples{
\dontrun{
central <- construct_central_dataframe(
   analysis_variable_df = .,
   dataframe_list = master,
   number.arms = N.Arms,
   id_cols = id_cols,
   field_df = fields,
   include.adverse = T,
   blinded = 'n')
}

}
