% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_baseline_table.R,
%   R/construct_screening_table.R
\name{construct_screening_table}
\alias{construct_screening_table}
\title{Creates the baseline characteristic table for reporting}
\usage{
construct_screening_table(
  trial.data,
  variable.details.df = variable.details.df
)

construct_screening_table(
  trial.data,
  variable.details.df = variable.details.df
)
}
\arguments{
\item{trial.data}{(dataframe) the dataframe containing all the trial data - ideally created by the data_construction function in this package.}

\item{variable.details.df}{(dataframe) the dataframe within the inputs folder which specifies which variables to be seen in the screening table - typically called variable.details.df}
}
\value{
(dataframe) returns a dataframe which summarises the baseline characteristics

(dataframe) returns a dataframe which summarises the screening characteristics
}
\description{
`construct_baseline_table` returns a dataframe which summarises the baseline characteristics per arm and overall
specified within the AnalysisVariablesDetails.xlsx. With mean, median, sd, IQR, min and max for numerical variables,
as well as counts of the number of each non-empty item in both numeric and categorical variables including each group within
the categorical variables.

`construct_screening_table` returns a dataframe which summarises the screening characteristics
specified within the AnalysisVariablesDetails.xlsx. With mean, median, sd, IQR, min and max for numerical variables,
as well as counts of the number of each non-empty item in both numeric and categorical variables including each group within
the categorical variables.
}
