% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont._summ_to_string.R
\name{cont.summ_to_string}
\alias{cont.summ_to_string}
\title{Create summaries from continuous data'}
\usage{
cont.summ_to_string(
  dataframe_object = .,
  continuous.data.name,
  population.list.obj = itt,
  include_mad = F,
  dec_places = 1,
  by_arm = T
)
}
\arguments{
\item{dataframe_object}{(dataframe) A dataframe you wish to summarise}

\item{continuous.data.name}{(string) Continuous variable which is a column name}

\item{population.list.obj}{(list) List object of the participant IDs for people
in each arm.}

\item{include_mad}{(boolean) Default FALSE. Whether to include the median
absolute deviation alongside the median and IQR}

\item{by_arm}{(boolean) Default TRUE. Whether to provide these summaries by
arm and overall}
}
\value{
(list)  Returns the descriptive summaries in the form of a list
which can be inputted straight into a summary table

 @examples age_summary<-cont.summ_to_string(core.dataframe,
                                            continuous.data.name = 'age_demographics')
}
\description{
`cont.summ_to_string` create descriptive summaries for continuous data.
This function is used in the summary table functions. It out puts the continuous
summaries as a list of lines you would see in a baseline summary table e.g.
mean (sd) or min, max etc
}
