% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_UTF8_known_filepath.R
\name{df_to_UTF8_known_filepath}
\alias{df_to_UTF8_known_filepath}
\title{Converts a dataframe into UTF-8 from a known filepath'}
\usage{
df_to_UTF8_known_filepath(df = ., file.path)
}
\arguments{
\item{df}{(dataframe) The dataframe to be converted}

\item{file.path}{(string) File path to location of the csv to be read in}
}
\value{
(dataframe)  Returns a dataframe from the file path with encoding
of UTF-8 instead the file's previous encoding.

 @examples df_to_UTF8_known_filepath(lookups,
                                     file.path = paste0(get_data, '/Lookups.csv'))
}
\description{
`df_to_UTF8_known_filepath`  Returns the dataframe with encoding
of UTF-8 instead the dataframe's previous encoding.
}
