% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing.R
\name{fill_missing}
\alias{fill_missing}
\title{Fill missing data at a timepoint with data from another specified timepoint'}
\usage{
fill_missing(
  multievent_data = .,
  focus_var,
  event_col = "event_name",
  target_event = "Baseline",
  replacement_event = "Screening"
)
}
\arguments{
\item{multievent_data}{(dataframe) A dataframe you wish to summarise}

\item{focus_var}{(string) Variable to have NAs filled}

\item{event_col}{(string) Column name of event names. Default 'event_name'}

\item{target_event}{(string) Event whose NAs to be replaced}

\item{replacement_event}{(string) Event who can replace NAs with available data}
}
\value{
(column)  Returns the more complete column

 @examples core_baseline<-fill_missing(core.dataframe,
                                       focus_var = 'age_demographics'
                                       )
}
\description{
`fill_missing` fills in NA values where possible from another specified
event
}
