% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_n_convert.R
\name{label_n_convert}
\alias{label_n_convert}
\title{Label columns descriptively and convert to the correct type'}
\usage{
label_n_convert(
  main.df = .,
  dupl_n = duplicated_names,
  label.dictionary = lookups,
  type.dictionary = fields
)
}
\arguments{
\item{main.df}{(dataframe) A dataframe you wish to add labels and convert the
column types. It is used within the construct_central_dataframe to provide
clarity on variables}

\item{dupl_n}{(string) Any duplicated column names}

\item{label.dictionary}{(dataframe) The data dictionary where the
label of the variable can be found}

\item{type.dictionary}{(dataframe) The data dictionary where the
type of the variable can be found}
}
\value{
(dataframe)  Returns the labeled and converted appropriately dataframe

 @examples core<-label_n_convert(unlabelled.core.dataframe)
}
\description{
`label_n_convert` labels and converts each column in the dataframe
according to its type in the data dictionary.
}
