% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_dict_duplicates.R
\name{mutate_dict_duplicates}
\alias{mutate_dict_duplicates}
\title{Add a description of the variable to the dataframe column'}
\usage{
mutate_dict_duplicates(dictionary_df = lookups, dupl_n = duplicated_names)
}
\arguments{
\item{dictionary_df}{(dataframe) The dataframe which needs the variable names
mutated in order to match with the central dataframe}

\item{dupl_n}{(vector) A vector of the duplicated names within the selected
CRFs to be mutated.}
}
\value{
(dataframe)  Returns the dataframe with the duplicated names amended
to distinct naming that follows the central dataframe nomenclature.

 @examples lookups<-mutate_dict_duplicates(dupl_n = c('age', 'sex', 'wd_dt'))
}
\description{
`mutate_dict_duplicates` adds a description (found in the specified data dictionary)
to the column for use in constructing the central dataframe. Adding descriptions
to each variable is an important part of data sharing and reproducibility.
}
