% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_nonempty.R
\name{N_nonempty}
\alias{N_nonempty}
\title{Returns the number of non-empty items in a vector'}
\usage{
N_nonempty(data.vector)
}
\arguments{
\item{data.vector}{(vector) a vector either numeric or other}
}
\value{
(numeric) returns the number of non-empty items
}
\description{
`rd_percent` returns the number of non-empty items in a vector
}
