% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical_summary_rounded.R
\name{numerical_summary_rounded}
\alias{numerical_summary_rounded}
\title{Returns summary statistics to specified decimal places '}
\usage{
numerical_summary_rounded(data, decimal_places = 1)
}
\arguments{
\item{data}{(vector) Vector of continuous data (can include NAs).}

\item{decimal_places}{(numeric) the number of decimal places}
}
\value{
(list) returns the summary object

 @examples numerical_summary_rounded(c(0, 1, 2, 3, 4, 5, NA))
}
\description{
`numerical_summary_rounded`  Returns summary statistics of numerical data,
 if the number of decimal places is not specified, the standard is one d.p.
}
