% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_summaries_perArmOverall.R
\name{percentage_summaries_perArmOverall}
\alias{percentage_summaries_perArmOverall}
\title{Returns percentages of items in each arm and overall'}
\usage{
percentage_summaries_perArmOverall(n, arm.n = N.Arms, itt = itt)
}
\arguments{
\item{n}{(vector) Vector of the number of people/items with a certain characteristic in each arm and overall. This should be inputted as c(control.n, intervention1.n, ..., overall.n)}

\item{arm.n}{(numeric) The number of arms in the study}

\item{itt}{(list) the list object of the participant IDs in each arm}
}
\value{
(vector) returns the list of percentages to one decimal place in the format c(control.percentage, intervention1.percentage, ..., overall.percentage)

 @examples percentage_summaries_perArmOverall(c(2, 5, 3, 10))
}
\description{
`percentage_summaries_perArmOverall`  Returns the percentage of people with a certain characteristic in each arm and overall to one decimal point'
}
