% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_percent.R
\name{rd_percent}
\alias{rd_percent}
\title{Returns the percentage of n in N to the nearest integer'}
\usage{
rd_percent(numerator.n, denominator.N)
}
\arguments{
\item{numerator.n}{(numeric) the number of items with a certain characteristic from the group}

\item{denominator.N}{(numeric) the number of items in the group}
}
\value{
(numeric) returns a percentage of n in N to the nearest integer

 @examples rd_percent(3, 10)
}
\description{
`rd_percent` returns the percentage of n in N to the nearest integer
}
