% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_to_UTF8.R
\name{read_csv_to_UTF8}
\alias{read_csv_to_UTF8}
\title{Reads in csv files to UTF-8'}
\usage{
read_csv_to_UTF8(file.path = ., ...)
}
\arguments{
\item{file.path}{(string) File path to location of the csv to be read in}
}
\value{
(dataframe)  Returns a dataframe from the file path with encoding
of UTF-8 instead the file's previous encoding.

 @examples read_csv_to_UTF8(paste0(get_data, '/Lookups.csv'))
}
\description{
`read_csv_to_UTF8`  Returns a dataframe from the file path with encoding
of UTF-8 instead the file's previous encoding.
}
