% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameCRF_toR.R
\name{renameCRF_toR}
\alias{renameCRF_toR}
\title{Converts a dataframe with CRF names into R compatible version'}
\usage{
renameCRF_toR(df = ., form.column.name = "form")
}
\arguments{
\item{df}{(dataframe) The dataframe to be converted}

\item{form.column.name}{(string) The column name containing the CRF form names}
}
\value{
(dataframe)  Returns a dataframe from the converted CRF names.

 @examples renameCRF_toR(fields,
                          form.column.name = 'Form')
}
\description{
`renameCRF_toR`  Returns the dataframe with R compatible converted CRF names.
                 This is particiularly helpful when reading things in from the
                  lookups file or any other data dictionary.
}
