% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_rounding_rm_na.R
\name{sd_rounding_rm_na}
\alias{sd_rounding_rm_na}
\title{Returns standard deviation (removes NAs) to one decimal point'}
\usage{
sd_rounding_rm_na(data)
}
\arguments{
\item{data}{(vector) Vector of continuous data (can include NAs).}
}
\value{
(numeric) returns the standard deviation to one decimal place

 @examples sd_rounding_rm_na(c(0, 1, 2, 3, 4, 5, NA))
}
\description{
`sd_rounding_rm_na`  Returns standard deviation to one decimal point and ignores NAs within the data
}
