% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_replacement_refInput.R
\name{suffix_replacement_refInput}
\alias{suffix_replacement_refInput}
\title{Extracts variables from specified data featuring a 'suffix'}
\usage{
suffix_replacement_refInput(column.name.vector, variable.details.df)
}
\arguments{
\item{column.name.vector}{(vector) Vector of variable names of a certain data
frame. One or more elements with contain 'suffix' as part of the variable
name. E.g.  c('consent_yn', 'no_cons_rsn_suffix',...)}

\item{variable.details.df}{(data frame) should be the data frame in the
Inputs folder. An example can be seen in the template folder.}
}
\value{
(vector) returns a vector with all variable names including the
  original non-'suffix' elements and all variable names within that data
  frame with the specified text inside the 'suffix' variable name. E.g.,
  c('consent_yn', 'no_cons_rsn_not_int', 'no_cons_rsn_not_avail', ...)
}
\description{
`suffix_replacement_refInput` returns the variable names which are standalone
  names and any names which correspond with the 'suffix' text.
}
\details{
'suffix' in a variable name denotes that whatever other text alongside
  suffix will be used to collect variable names in the specified data frame
  For example 'suffix_fa' will collect all variable names which include _fa
  This is used when inputting variables from the
  AnalysisVariablesDetails.xlsx which names the variables to be in the master
  data frame.
}
\examples{
\dontrun{
df_spec_cols<-suffix_replacement_refInput(df_spec_cols)
}
}
