% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table_by_arm.R,
%   R/summary_table_overall.R
\name{summary_table_by_arm}
\alias{summary_table_by_arm}
\title{Creates a summary table by arm and overall for reporting}
\usage{
summary_table_by_arm(
  data_to_summarise = .,
  summary_variables = analysis_variable_df$`Screening-Summary-Variables`,
  population.list.obj = itt,
  id_cols = c("screening", "event_name")
)

summary_table_by_arm(
  data_to_summarise = .,
  summary_variables = analysis_variable_df$`Screening-Summary-Variables`,
  population.list.obj = itt,
  id_cols = c("screening", "event_name")
)
}
\arguments{
\item{data_to_summarise}{(dataframe) the dataframe containing all the trial data -
ideally created by the construct_central_dataframe function in this package.}

\item{summary_variables}{(vector) the variables to be summarised within this table}

\item{population.list.obj}{(list) a list object of the same length as the
  number of arms. Each list element is a character vector containing the
  screening numbers of participants in an arm in the relevant analysis
  population.
#'}

\item{id_cols}{(vector) the names of the id columns as a character vector.
Default is `c("screening", "event_name", "event_id")`.}
}
\value{
(dataframe) returns a dataframe which summarises the specified
   characteristics by arm and overall

(dataframe) returns a dataframe which summarises the specified
   characteristics by arm and overall
}
\description{
`summary_table_by_arm` returns a dataframe which summarises the
  specified characteristics per arm and overall. With mean, median, sd,
  IQR, min and max for numerical variables, as well as counts of the number
  of each non-empty item in both numeric and categorical variables
  including each group within the categorical variables. There is also an
  option to add the median absolute deviation to the continuous summaries.

`summary_table_overall` returns a dataframe which summarises the
  specified characteristics overall only with no percentages. With
  mean, median, sd, IQR, min and max for numerical variables, as well as counts of the number
  of each non-empty item in both numeric and categorical variables
  including each group within the categorical variables. There is also an
  option to add the median absolute deviation to the continuous summaries.
  This function was made with the screening data in mind as this table is
  overall only and does not provide percentages because not all screened
  participants will answer each question e.g. reason not consented
}
