% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniq_nonNA.R
\name{uniq_nonNA}
\alias{uniq_nonNA}
\title{Finding unique non-NA elements}
\usage{
uniq_nonNA(item.list)
}
\arguments{
\item{item.list}{(vector) vector of items which can include NAs}
}
\value{
(vector) a vector of unique objects within the inputed 'item.list'

 @examples master.df<-construct_analyses_dataframe(variable.details.df, blinded='y', "identification_log")
}
\description{
`uniq_nonNA` returns a list of unique items within a list. These unique items
will not include NAs. This function is primarly used for identifying unique IDs within
the filtered segments of the master df created by the construct_master_dataframe
}
