% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_master_dataframe.R
\name{construct_master_dataframe}
\alias{construct_master_dataframe}
\title{Master dataframe construction}
\usage{
construct_master_dataframe(
  variable.details.df,
  blinded = "y",
  name.of.screening.df
)
}
\arguments{
\item{variable.details.df}{(dataframe) should be the dataframe in the Inputs folder.
An example can be seen in the template folder}

\item{blinded}{(character) This should be y or n}

\item{name.of.screening.df}{(string) This is the name of the PROSPECT csv file which
contains the details of all participants screened. In some trials, this is the
"identification_log" or simply "screening_log".}
}
\value{
(dataframe) master dataframe to be used for analyses or as a basis
 for sharing

 @examples master.df<-construct_master_dataframe(variable.details.df, blinded='y', "identification_log")
}
\description{
`construct_master_dataframe` returns a master dataframe which can be used
for primary analyses, or as a base dataset for sharing .
}
\details{
From the dataframes read in by Read-data.R, this function will produce a
dataframe of all participants (from screening onwards) with the corresponding
variables as specified in Inputs/AnalysisVariablesDetails.xlsx.

Any constant characteristics for each participant will be displayed in its own
column, for example: sex, ethnicity

Any date variables will need to be manipualted if another format other than the PROSPECT format is needed.

Adverse events will also be included to this dataset (if specified within
the input xlsx)

If blinded='y', a dummy randomisation will be implemented. If blinded='n',
the randomisation allocation will be taken from PROSPECT.
}
